app/services/content_service.py
from typing import Dict, Any, List
from openai import AsyncOpenAI
from dotenv import load_dotenv
import os
import json

load_dotenv()

client = AsyncOpenAI(api_key=os.getenv("OPENAI_API_KEY"))

class ContentService:
    @staticmethod
    async def generate_sample_content(components: List[Dict[str, Any]], context: str) -> Dict[str, Any]:
        """Generate sample content for UI components."""
        try:
            # Return empty content for now
            return {}
        except Exception as e:
            print(f"Error generating content: {str(e)}")
            return {}

    @staticmethod
    def get_placeholder_images(components: List[Dict[str, Any]]) -> Dict[str, str]:
        """Get placeholder images for components that need them."""
        try:
            # Return empty images for now
            return {}
        except Exception as e:
            print(f"Error getting placeholder images: {str(e)}")
            return {}
