requirements.txt
apify-client>=1.4.0
anthropic>=0.7.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-dotenv>=1.0.0
aiohttp>=3.9.0
jinja2>=3.1.0
pydantic>=2.5.2
openai>=1.3.7
python-multipart>=0.0.6
Pillow>=10.1.0
requests>=2.31.0
